Add ML2 settings; return proper value for neutron_tenant_network_type
This commit adds a number of parameters required for ML2 compatibility;
it also ensures that $neutron_tenant_network_type is set correctly. If
ML2 is in use, $neutron_tenant_network_type needs to be an array. If the
old OVS plugin is in use, it needs to be a string.