Add database configuration to neutron.conf
Previous versions of Kickstack would have the database configuration in
the plugin. By now, apparently, the recommended mechanism is to have the
DB configuration happen in neutron.conf -- at least the ML2 class of the
Neutron module doesn't even implement database parameters anymore. Thus,
we need the database configuration in neutron.conf. This commit adds the
required parameters to the neutron class.