Add ML2 parameters to the kickstack::neutron::plugin class
This commit enables the kickstack::neutron::plugin class to support the
ML2 plugin in OpenStack Neutron. It ensures that network_vlan_ranges and
tunnel_id_ranges are set properly depending on the plug-in that is used.
For OVS or Linuxbridge setups, these need to be strings. In ML2 setups,
they need to be arrays.

It also deletes a workaround for a previous bug in the Neutron plugin
for Open vSwitch that avoided proper functioning of the firewall_driver
variable. The bug in the Neutron plugin has been fixed by now, and thus
the workaround is not necessary anymore.