Add the auth_encryption_key parameter to the heat::engine call
More recent versions of the puppet-heat module, starting in Dec '13,
require a auth_encryption_key parameter to be passed to the engine
class. Kickstack didn't pass on the parameter, which caused the
installation of Heat to fail. This commit adds the parameter, sets
it automatically and exports it as fact.

Applying this fix makes deploying Heat via Kickstack possible again.